+++ identity-apps-core/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @wso2is/identity-apps-core
 
+## 5.3.2
+
+### Patch Changes
+
+- [#10455](https://github.com/wso2/identity-apps/pull/10455) [`09983a3bc71da637500f30a53ca852ec7fba42b9`](https://github.com/wso2/identity-apps/commit/09983a3bc71da637500f30a53ca852ec7fba42b9) Thanks [@DilshanSenarath](https://github.com/DilshanSenarath)! - Stop displaying the resend success message when the resend operation fails
+
+## 5.3.1
+
+### Patch Changes
+
+- [#10447](https://github.com/wso2/identity-apps/pull/10447) [`7a47be8939aa0984e35f100b84bb7230e41fd7dd`](https://github.com/wso2/identity-apps/commit/7a47be8939aa0984e35f100b84bb7230e41fd7dd) Thanks [@NipuniBhagya](https://github.com/NipuniBhagya)! - Fix localization issue with product footer
+
+## 5.3.0
+
+### Minor Changes
+
+- [#10437](https://github.com/wso2/identity-apps/pull/10437) [`7b3c52b42da27068217c63ceada4336e661dddc9`](https://github.com/wso2/identity-apps/commit/7b3c52b42da27068217c63ceada4336e661dddc9) Thanks [@DilshanSenarath](https://github.com/DilshanSenarath)! - Add provider-level OTP delivery failure notifications for Email OTP and SMS OTP
+
 ## 5.2.5
 
 ### Patch Changes